YAML to proto converter (#22)
Adds support for reading yaml and converting to one of the proto messages defined in the [api](https://github.com/envoyproxy/xds-relay/tree/master/api/protos/aggregation/v1).

We convert from yaml to proto via an intermediary JSON representation. Note that we are using the [new version](https://blog.golang.org/a-new-go-api-for-protocol-buffers) of the golang API for protocol buffers. We're leveraging the [protojson](https://pkg.go.dev/google.golang.org/protobuf/encoding/protojson?tab=doc) package to handle the unmarshaling from JSON to the actual protos, more specifically so that `oneof` fields are handled correctly. It's also worth pointing out that the generated code for the protos is actually readable in this new version of the API.

Of note I also fixed a mistake in the initial definition of the protos, more specifically, how we were defining a conjunction of `ResultPredicate` was convoluted. I took the opportunity to rename the field from `RepeatedResultPredicate` to `AndResult`. 

Issue: https://github.com/envoyproxy/xds-relay/issues/6

Signed-off-by: eapolinario <eapolinario@lyft.com>