Don't extend TTL for AddWatch (#60)
Fixes #58 

The expiration time should only be delayed in SetResponse, since the intent of the TTL was to avoid stale responses in the cache. Adding/removing watches should not affect the expiration time of a key. This change removes that extension from AddWatch, and #59 already omits the expiry time extension.

Signed-off-by: Lisa Lu <lisalu@lyft.com>