DeleteRequest API (#59)
Fixes #57

This change adds an API to the cache to support deleting a request when given the key and the (pointer to the) request, which the orchestrator can use upon cancellation. If the key is not in the cache, nothing happens.

The input type to AddRequest was also changed to take a pointer to a request.

Signed-off-by: Lisa Lu <lisalu@lyft.com>