End-to-end tests (#62)
This PR adds support for end-to-end tests involving `go-control-plane` SnapshotCache as the backbone for a management server, `xds-relay` server and one instance of envoy.

The setup is similar to how `go-control-plane` does its [e2e tests](https://github.com/envoyproxy/go-control-plane/blob/master/pkg/test/main/main.go), although in our case we decided to adopt a more conventional test structure, the idea being that we're going to extend the number of tests as time goes by.

Speaking about the mechanics of the tests, we run them in a container to make it easy to bring a real instance of envoy. By choosing to do it all in golang we had to sacrifice portability since there is no cross-platform way of safely exiting processes spawned in go. I ran multiple tests with more than one test in the same suite binding to the same ports (basically duplicates of the tests, only with different names) and had to introduce a 1 second delay between tests to make sure the OS can free up the ports between tests.

Signed-off-by: eapolinario <eapolinario@lyft.com>