Explicitly close downstream watches
Although the language specification for Go specifies that channels will
be garbage collected when no longer referenced, the receiver (
go-control-plane in this scenario) may be still looking for an explicit
signal. In our experiments, we saw that there were scenarios where cache
TTL triggering removal of downstream watchers, did not cause the Envoy
clients to resend requests.

Signed-off-by: Jess Yuen <jyuen@lyft.com>