Remove redirect behavior for admin endpoints and support getting log level (#129)
This change:

- Supports returning the current log level in string format when no parameter is supplied to the admin logging endpoint.
- Handles trailing slashes in URLs, such that paths with and without are registered to the same handlers without any redirects. [The redirects are a characteristic of http handlers in Go](https://golang.org/pkg/net/http/#ServeMux):
```If a subtree has been registered and a request is received naming the subtree root without its trailing slash, ServeMux redirects that request to the subtree root (adding the trailing slash). This behavior can be overridden with a separate registration for the path without the trailing slash. For example, registering "/images/" causes ServeMux to redirect a request for "/images" to "/images/", unless "/images" has been registered separately.```

Fixes #127 

Signed-off-by: Lisa Lu <lisalu@lyft.com>