Improve integration test (#142)
The integration tests were not stopping envoy between each version.
The test checks that the hello world http server is able to respond. Since envoy was not restarted between versions, the tests were still using v2 configurations and passing.

Also we've identified the reason of false positive. We were not checking the output after starting envoy, in case envoy never starts, the buffer is empty.

Signed-off-by: Jyoti Mahapatra jmahapatra@lyft.com