Evict cache on upstream failure (#146)
When upstream control plane rotates, the grpc client has to make a new connection. Today we recognize an issue in the stream and dont retry. We should at least close the downstream sidecar streams, so that we get some semblance of retry from envoy sidecars. This could cause thundering herd, but this is a stopgap until the next PR fixes the retry.

Signed-off-by: Jyoti Mahapatra jmahapatra@lyft.com