Fix flakey tests (#163)
In the upstream client tests, the defer way of closing goroutines was turning out to be flakey. By the time the go-leak ran the goroutines were not always done.
Instead we have to explicitly wait for each test cleanup before the go-leak ran. In go tests there is no way to add a per test teardown hook.

Verfied each test with -count=100

Signed-off-by: Jyoti Mahapatra jmahapatra@lyft.com