Add ready endpoint (#184)
While rolling out in staging we realized that there are sometimes cx issues due to which there is cache mismatch between upstream and xdsrelay. Due to the severity of traffic we end up terminating the erroneous pod. We need a way to web off faulty pods to continue investigating the issue.

Signed-off-by: Jyoti Mahapatra jmahapatra@lyft.com
Signed-off-by: Jyoti Mahapatra <jmahapatra@lyft.com>